requirements.txt
dj-database-url==0.5.0
Django==3.2.6
django-admin-interface==0.16.3
django-ckeditor==6.1.0
django-colorfield==0.4.2
django-crispy-forms==1.12.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
sqlparse==0.4.1
psycopg2-binary==2.9.1
django-heroku==0.3.1